Feature -  Non-Blurry Switches and Checks (#277)
* Use pre-rendered assets to ensure sharpness

GTK CSS seems to render SVG assets to PNG prior to scaling them. This results in a blurry
image for the switches, checkboxes, and radio buttons on HiDPI. Instead, we should
pre-render these images at 1x, 2x, and 3x sizes to ensure they are always sharp.

This allows us to have sharp images at all sizes, and to avoid depending on the
icon theme to provide these images (icons are also always sharp). Otherwise there would
be no images if a user was using the gtk theme with a different icon theme.

* Update changelog