Big change. This makes Pop more upstream-compliant
This is the result of making sure Pop does not break upstream GNOME apps. It's very
Pop-y, but with metrics and features that give it 100% feature parity with Adwaita
which helps ensure upstream apps will not break if a user is using this theme. It's
current incomplete, so it should be considered alpha-quailty for now.