Upstream compliant theming (#322)
* Big change. This makes Pop more upstream-compliant

This is the result of making sure Pop does not break upstream GNOME apps. It's very
Pop-y, but with metrics and features that give it 100% feature parity with Adwaita
which helps ensure upstream apps will not break if a user is using this theme. It's
current incomplete, so it should be considered alpha-quailty for now.

* Install dark theme too

* Render assets, style scales

* Add elementary widgets styling

* Fix Search results width

* Ignore build obj folder

* Calendar fixes

* Menu fixes

* Use dark theme for GDM

* Fix broken GResource

* Remove screenshots from README

These have been outdated for a long time anyway, and don't represent an accurate
image of the theme. Lots of images of the theme are available in screenshots online,
and these represent an undue maintenance burden.

Also point the logo towards master

* Fix broken GResource for light theme too.

* Shell: Calendar day fixes

* Shell: Contrast improvement on close button

* Fix large user icon in system menu

* Fix lintian error with changelog

* Documentation updates

* Move slider styling out of _common and into Tweaks for easier maintenance

* Improve dark theme selected item backdrop foreground color

* Fix borders on Nautilus list-view headers

* Correct typo with color pallete

* Fix old color functions and replace with rgba

* Fix path issue with assets

* Attempt to fix upgrade errors with debian packaging.

* Fix theme installation error.

* Make sure we're hitting the correct GNOME Shell theme version

* Remove Extra files.