Rename source package back to pop-gtk-theme (#330)
This is not ideal from a heuristics perspective, because it's not just
a gtk-theme, but also a theme for GNOME Shell. It also signifies our
intent to be taken seriously within the GNOME ecosystem. However, it
breaks some packaging, so...

FORCE MERGE: Does not affect delivered software, only the internal build system.