build(GTK): Fix installation of gtk-3.0 folders.
When we copied the contents of the gtk-3.20 folder into the gtk-3.0 folder,
this also copied the build configuration. This meant that the (identical)
sources in the 3.0 folder were being installed into the gtk-3.20 folder,
and no files were being installed to the gtk-3.0 folder. This corects
this mistake.

Fixes #447