fix(apps): Tweaks to improve the Firefox tab line color in Hirsute (#515)
* fix(apps): Tweaks to improve the Firefox tab line color

This unfortunately affects all of Firefox, so we also darken the color so enable
it to have somewhat acceptable contrast. It is confined to Firefox though.

* fix(apps): Firefox: new selection bg color

This sets the selection background color in Firefox to the same color in both
themes. It's currently lighter in the light theme, and darker in the dark theme. While
contrast is iffy with this set up in the light theme, it's much better in the dark
theme and we use this combination elsewhere in the theme. Hopefully the Firefox
bug is fixed soon.