fix(panel): use border for margin
This ensures that the button widget extends to the screen edge, allowing clicks
from the edge of the screen. This might look weird if someone uses an extension
to make the panel transparent as there will be a black border around the button,
but it looks correct with the default theming.